% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jellyforecast-package.R
\docType{package}
\name{jellyforecast}
\alias{jellyforecast-package}
\alias{jellyforecast}
\title{jellyforecast}
\description{
Serves a subset of ecopmo jellyfish predictive maps
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/jellyforecast}
}

}
\keyword{internal}
