% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot a forecast}
\usage{
plot_forecast(
  cfg,
  x = read_raster(cfg),
  wrap = length(dim(x)) > 2,
  crop = NULL
)
}
\arguments{
\item{cfg}{configuration list}

\item{x}{stars object}

\item{wrap}{logical, if TRUE create a facet wrapped single image, otherwise
a list of ggplot objects are returned.  If only one time is provided in
x then this ignored and a list is returned}

\item{crop}{NULL or bbox to crop the data}
}
\value{
either a single ggplot object (facet wrapped by time) or
a list of ggplot objects (one per unit of time)
}
\description{
Plot a forecast
}
