% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_coastline}
\alias{read_coastline}
\alias{read_cities}
\title{Read a coastline or cities}
\usage{
read_coastline(path = package_path())

read_cities(path = package_path())
}
\arguments{
\item{path}{chr data path}
}
\value{
sfc (geometry) object
}
\description{
Coastline made with Natural Earth. Free vector and raster map data at naturalearthdata.com.
}
