% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_products}
\alias{read_products}
\title{Read the product table}
\usage{
read_products(
  path = package_path(),
  extra = c("daily", "wrapped", "data")[1:2]
)
}
\arguments{
\item{path}{chr data path}

\item{extra}{chr, vector of extras to add as columns such as "daily"
and "wrapped" images as well as raw "data" (default "daily", "wrapped")}
}
\value{
data frame with species, version, cfg and possibly daily, wrapped and data columns
}
\description{
Read the product table
}
