% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_graphics}
\alias{save_graphics}
\title{Save graphics as PNGs}
\usage{
save_graphics(cfg, x = plot_forecast(cfg), path = data_path(), wipe = TRUE)
}
\arguments{
\item{x}{either a ggplot object (facte wrapped by time) or a list
of daily graphics}

\item{path}{the path to write to}

\item{wipe}{logical, when writing daily files, wipe the existing ones first?}
}
\value{
the input object
}
\description{
Save graphics as PNGs
}
