% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_config}
\alias{write_config}
\alias{read_config}
\alias{list_configs}
\title{Read, write and list configuration lists}
\usage{
write_config(x, path = data_path())

read_config(x, path = package_path())

list_configs(path = package_path())
}
\arguments{
\item{x}{configuration list}

\item{path}{chr, data path}
}
\value{
for \code{write_config} and \code{read_config} a configuration list
}
\description{
Read, write and list configuration lists
}
