% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_raster}
\alias{write_raster}
\alias{read_raster}
\title{Read and write raster data}
\usage{
write_raster(x, cfg, path = data_path())

read_raster(cfg, path = package_path())
}
\arguments{
\item{x}{stars object}

\item{cfg}{configuration list}

\item{path}{chr data path}
}
\value{
for \code{write_raster} and \code{read_raster} a stars object
}
\description{
Read and write raster data
}
